restructure xcat files
