+++ provision/roles/provision_validation/tasks/validate_admin_nic.yml
