+++ provision/roles/provision_validation/tasks/validate_bmc_nic.yml
