+++ provision/roles/provision_validation/tasks/validate_disk_partition_vars.yml
