+++ provision/roles/provision_validation/tasks/validate_mapping_file.yml
