+++ provision/roles/provision_validation/tasks/validate_public_nic.yml
