+++ provision/roles/provision_validation/tasks/validate_pxe_nic.yml
