+++ provision/roles/xcat/files/utils/mapping_ib.py
