+++ provision/roles/xcat/tasks/configuration/configure_xcat.yml
